pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools-scm[toml]>=4.2"]
build-backend = "setuptools.build_meta"

[project]
name = "cardtool"
authors = [
    {name = "Daniel Aucatoma", email = "d.aucatoma96@gmail.com"}
]
description = "An encrypted card generator tool"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["DUKPT", "credit", "card", "tool"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "click==8.1.3",
    "pycryptodome==3.18.0",
    "PyYAML==6.0",
    "loguru==0.7.0",
    "jsonschema==4.17.3",
    "pydash==7.0.4",
    "ber-tlv==0.0.6",
    "simplejson==3.19.1",
    "dill==0.3.6",
    "toolz==0.12.0"
]
dynamic = ["version"]

[project.scripts]
cardtool = 'cardtool.command.main_group:cli_card'

[tool.setuptools_scm]

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["**/__init__.py"]

[tool.coverage.report]
precision = 2
skip_covered = false
fail_under = 100

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
testpaths = [
    "tests",
]
norecursedirs = "tests/helper"
addopts = [
    "--import-mode=importlib"
]

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.autoflake]
remove-unused-variables = true
remove-all-unused-imports = true